Add a pixmap queue for displaying the world.
Now there are two spinners - the GUI spinner and the world spinner which
enqueues pixmaps for the GUI spinner to show. The pixmap queue will hold
up to 300 turns in it. (roughly 30 seconds at 100ms per display).

This does have some anti-aliasing issues because we need to draw the map into
a pixbuf and then scale it up / down before displaying.