Implement unit collisions for enemy units. Remove melee attack.
Now, if enemy units are on the same square, the team with the most units
survives and loses the amount of units that the next largest team on
that square owns.

For example, if A has 10 units, B has 6 and C has 3 and they are all on
one square, B and C both lose all their units while A loses 6 units.